Help users who have JS disabled
When the user loads the viewer page that requires JS to function
correctly without JS enabled, they are told that the page requires JS
and that they either have to enable it or they can use the lightweight
version that does not require it.

Previously, this was missing and therefore a user that did not have
JavaScript enabled on their browser would think that there was no way
to use the application.

I also added a link to a new page that links to all of the individual
lightweight HTML pages because the user is not expected to know the
URLs.