Added icon
Added an SVG icon made with Inkscape to the repo. It is converted into a
favicon automatically by the build file and also converted to the format
read by Safari on iOS. I edited the PWA manifest file to add a reference
to that icon. The PWA is still not installable because the app is
missing a service worker.