Change theme-color to white
The application can be either in light or dark mode. Unfortunately, I
don't know a clean way to set the theme-color dependending on the
current mode so I previously decided to set it to black to match the
dark mode. This means that when light mode is enabled, the app looks
bad.

However, it seems that Chrome on Android only respects the theme-color
value when Chrome is in light mode, which means that we should set the
value to white so that when Chrome is in light mode, it shows the white
bar and when it is in dark mode, it shows the default dark bar.